+++ R/paths.R
@@ -330,6 +330,7 @@ ggdag_paths <- function(
   if (missing(edge_type)) {
     edge_type <- ggdag_option("edge_type", "link_arc")
   }
+  edge_type <- check_edge_type(edge_type)
   edge_engine <- match.arg(edge_engine, c("ggraph", "ggarrow"))
 
   p <- if_not_tidy_daggity(.tdy_dag, ...) |>